Code clean-up, fixed missing license headers, moved classes in their correct folders.
