+++ src/matcracker/BedcoreProtect/commands/BCPCommand.php
@@ -24,9 +24,9 @@ namespace matcracker\BedcoreProtect\commands;
 use BlockHorizons\BlockSniper\sessions\SessionManager;
 use matcracker\BedcoreProtect\Inspector;
 use matcracker\BedcoreProtect\Main;
+use matcracker\BedcoreProtect\math\Area;
+use matcracker\BedcoreProtect\math\MathUtils;
 use matcracker\BedcoreProtect\ui\Forms;
-use matcracker\BedcoreProtect\utils\Area;
-use matcracker\BedcoreProtect\utils\MathUtils;
 use matcracker\BedcoreProtect\utils\Utils;
 use pocketmine\command\Command;
 use pocketmine\command\CommandSender;
