+++ src/matcracker/BedcoreProtect/enums/Action.php
@@ -19,7 +19,7 @@
 
 declare(strict_types=1);
 
-namespace matcracker\BedcoreProtect\utils;
+namespace matcracker\BedcoreProtect\enums;
 
 use InvalidArgumentException;
