+++ src/matcracker/BedcoreProtect/enums/RegistryTrait.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace matcracker\BedcoreProtect\utils;
+namespace matcracker\BedcoreProtect\enums;
 
 use ArgumentCountError;
 use Error;
