+++ src/matcracker/BedcoreProtect/listeners/BlockListener.php
@@ -21,8 +21,8 @@ declare(strict_types=1);
 
 namespace matcracker\BedcoreProtect\listeners;
 
+use matcracker\BedcoreProtect\enums\Action;
 use matcracker\BedcoreProtect\Inspector;
-use matcracker\BedcoreProtect\utils\Action;
 use matcracker\BedcoreProtect\utils\BlockUtils;
 use pocketmine\block\Bed;
 use pocketmine\block\Block;
