+++ src/matcracker/BedcoreProtect/storage/queries/QueriesEntitiesTrait.php
@@ -22,8 +22,8 @@ declare(strict_types=1);
 namespace matcracker\BedcoreProtect\storage\queries;
 
 use matcracker\BedcoreProtect\commands\CommandParser;
-use matcracker\BedcoreProtect\utils\Action;
-use matcracker\BedcoreProtect\utils\Area;
+use matcracker\BedcoreProtect\enums\Action;
+use matcracker\BedcoreProtect\math\Area;
 use matcracker\BedcoreProtect\utils\Utils;
 use pocketmine\entity\Entity;
 use pocketmine\Player;
