+++ src/matcracker/BedcoreProtect/storage/queries/QueriesInventoriesTrait.php
@@ -22,11 +22,11 @@ declare(strict_types=1);
 namespace matcracker\BedcoreProtect\storage\queries;
 
 use matcracker\BedcoreProtect\commands\CommandParser;
+use matcracker\BedcoreProtect\enums\Action;
+use matcracker\BedcoreProtect\math\Area;
+use matcracker\BedcoreProtect\primitive\PrimitiveBlock;
 use matcracker\BedcoreProtect\tasks\async\AsyncInventoriesQueryGenerator;
 use matcracker\BedcoreProtect\tasks\async\AsyncLogsQueryGenerator;
-use matcracker\BedcoreProtect\utils\Action;
-use matcracker\BedcoreProtect\utils\Area;
-use matcracker\BedcoreProtect\utils\PrimitiveBlock;
 use matcracker\BedcoreProtect\utils\Utils;
 use pocketmine\inventory\ContainerInventory;
 use pocketmine\inventory\Inventory;
